backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0
email-validator==2.1.0
pydantic-settings==2.1.0
itsdangerous==2.1.2

# Caching
redis==5.0.1
hiredis==2.2.3

# HTTP Client
httpx==0.25.2

# AWS Services
boto3==1.34.14
botocore==1.34.14

# AI & Text Processing
openai==1.3.7
beautifulsoup4==4.12.2
trafilatura==1.6.3
newspaper3k==0.2.8
lxml==4.9.3

# PDF Generation
reportlab==4.0.7
Pillow==10.1.0
pydantic==2.5.2
PyPDF2==3.0.1

# Fax Service
requests==2.31.0

# Task Queue
celery==5.3.4

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
aiofiles==23.2.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring
sentry-sdk[fastapi]==1.39.1